requirements.txt
Flask==2.3.3
Authlib==1.2.1
python-dotenv==1.0.0
requests==2.3.1
filelock==3.12.2
